Fastapi

+++ Week-3/Day-1/TodoApp/routers/todos.py
@@ -6,6 +6,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from starlette import status
 from pydantic import BaseModel,Field
+from routers.auth import get_current_user
 
 
 
@@ -22,6 +23,11 @@ def get_db():
 
 
 db_dependency=Annotated[Session,Depends(get_db)]
+user_dependency= Annotated[dict,Depends(get_current_user)]
+
+
+
+
 
 class TodoRequest(BaseModel):
     title:str=Field(min_length=3)
@@ -31,12 +37,17 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db:Annotated[Session,Depends(get_db)]):
-    return db.query(Todos).all()
+async def read_all(user:user_dependency,db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id==user.get('User Id')).all()
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def read_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model=db.query(Todos).filter(Todos.id==todo_id)\
+    .filter(Todos.owner_id==user.get('User Id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404,detail="Todo not found")
@@ -44,19 +55,22 @@ async def read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
 
 
 @router.post("/todo",status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency,todo_request:TodoRequest):
-    todo_model=Todos(**todo_request.model_dump())
+async def create_todo(user:user_dependency,db:db_dependency,todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model=Todos(**todo_request.model_dump(),owner_id=user.get('User Id'))
     db.add(todo_model)
     db.commit()
     db.refresh(todo_model)
     return todo_model
 
 @router.put("/todo/{todo_id}")
-async def update_todo(db:db_dependency,
+async def update_todo(user:user_dependency,db:db_dependency,
                       todo_id:int,
                       todo_request:TodoRequest):
     
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+    todo_model=db.query(Todos).filter(Todos.id==todo_id)\
+        .filter(Todos.owner_id==user.get('User Id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
     todo_model.title=todo_request.title
@@ -68,10 +82,12 @@ async def update_todo(db:db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def delete_todo(user:user_dependency,db:db_dependency,todo_id:int):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model=db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('id')).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404,detail="id not found")
-    db.query(Todos).filter(Todos.id==todo_id).delete()
+    db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('id')).delete()
     db.commit()
